feat: first prototype of low-level event-based parsing, heavily influenced by @braingram's prototype
I like this approach in analogy to the design of YAML parsers
themselves.  This is a lower-level interface that most users will not
interact with, though we'll include parts of it in the public API spec.

Includes a debug command for parsing an ASDF file and printing events as
they're encountered.  Still has bugs; not working quite right yet.