fix: correctly end parsing in an error state if YAML parsing failed
NOTE: We *could* in principle still try to recover even if we reached
invalid YAML--for example it might still be possible to recover the
block data.

Future work could include marking which types of errors are potentially
recoverable.  For example, a missing #ASDF header isn't the end of the
world; can fall back on the most recently supported standard.