refactor: rewrite the asdf_info function using just asdf_ API, no direct libfyaml
This turns out to be quite onerous, because in order to print a tree in
this format one *has* to build a document tree first in order to know
the depth and number of children of each node.  So we end up
reimplementing (badly) libfyaml's own code for building the document
tree.

Nevertheless, it was a good excerise of the low-level asdf_event API for
now.  Will likely go through another rewrite (more resembling the
original version) once we add asdf_ specific wrappers around the
document tree.