refactor: stream generalization
Different stream types are supportable, currently FILE* and memory
buffer based, the two most likely to be common.

Streams are either seekable or non-seekable, though most of the parsing
logic assumes non-seekable streams (even though it's a corner case).

This is based on a next()/consume() model: stream->next() returns a
pointer to the next bytes available in the stream, as well as how many
bytes are available.  In the case of, say, the full file in memory then
the available bytes is just the full memory minus a position in the
buffer.  In the case of a file stream it may be less, as files are
buffered into memory a page at a time.

stream->consume() marks some number of bytes as consumed and advances
the stream to the next available bytes.

This works well for the case of non-seekable streams, as well as keeps
the logic clean for seekable streams.

In the case of seekable streams we can still seek around the file as
necessary, for example to find a block index, or return to the location
of a block.  This capability isn't used yet but will be later.  In the
case of blocks we have to have the option, when generating a BLOCK
event, to either capture the full block data in a memory buffer, or skip
past it (or maybe even other options like tee it to a temp file).  The
present architecture makes many options possible depending on the use
case.

gh-16