fix: require a minimum number of bytes available when calling asdf_stream_next
This is particularly necessary when parsing the block header to ensure
that if some field is straddling the end of the read buffer, we refill
the buffer to have enough bytes available.

An earlier draft of this code actually had this feature but I removed
it under a misguided belief that it wasn't necessary (in part because
my tests happened to pass without it).

gh-16