refactor: implement sharing of asdf_context_t with the stream objects
now streams don't have to have their own ad-hoc level of error handling;
error context is shared between the stream and the parser (and other
bits in the future).

Use of stdatomic primitives for reference counting is strictly optional
and not really needed since for now we don't intend to make the parser
multi-threaded.  But I thought it would be fun, and it provides a little
future-proofing :)

gh-5