feat: support the top-level core/asdf-1.1.0 schema gh-38
refactor: disable for deallocation of extension objects when the value
they were read from is destroyed, this led to way too many foot-guns.

Unfortunately user is now responsible for freeing those objects.  This
is less than ideal, but can be fixed as part of gh-34 (probably add
reference counting to extension objects).