feat: all the *bare minimal* bits to parse a GWCS with FITS WCS transform
This is very flaky!  It doesn't know about all the different tag
versions and is tuned just to work with the one example file I am
testing against.  It really brings to light the need for better tag
version management, so OK we'll focus on that.

Also make distcheck is broken; will look into that in a follow-up

gh-66