build: for now remove the configure flag for wcslib support
In the end it was not really needed--there is no obvious advantage *for
now* in the gwcs extension providing a method to return a pre-populated
wcsprm.  Instead it just provides all the information needed, and can be
left independent of the wcs implementation

(e.g. my SourceXtractor++ just uses the structures provided here to
populate the wcsprm, since it actually *uses* wcslib to perform
transforms, and the setup is very straightfoward)

gh-67